Fix the lock implementation.
The acquire methods need to spin only when compareAndSet fails. They can return immediately when the lock cannot be acquired. To make this work, a shared release must balance each shared acquisition, because AbstractQueuedSynchronizer wakes only one waiting thread on release. Before, this change, the acquisition was not balanced by a release, so the acquisition needed to spin in every case, and that was very expensive. After this change, there is a major improvement in performance, especially under high contention.