Mix random reads with the writes.
Each transfer reads two accounts and then writes them from the same thread. With this access pattern the synchronized ArrayList (mutex vector in C++) looks very good. This change mixes random reads with the writes to simulate a workload that's read heavy. Under this different access pattern the CAS list (same in C++) looks very good.  The difference occurs, because the use of synchronized (mutex) turns a read into a write of the lock so the value must be owned by a core, whereas with CAS a read is only a read so the value can be shared among cores.