Move the logical lock into the shard (C++).
This allows the logical lock to share the cache-line. Before this change, acquiring the logical lock would synchronize one cache-line, and then acquiring mutex on the shard would synchronize another. After this change, acquiring the logical lock synchronizes that cache-line, and then the header for the shard is right there with it.